+++ nodemon.json
@@ -1,0 +1,4 @@
+{
+    "ignoreRoot": [".git"],
+    "restartable": false
+}
